Move test trace on str variables to variables test file
